Renamed strategy to push-voting-power
